requirements.txt
playwright>=1.48
httpx>=0.27
pydantic>=2.8
python-dotenv>=1.0
pillow>=10.4
